requirements.txt
# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0

# Configuration & Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
tqdm>=4.65.0

# Visualization & Monitoring
matplotlib>=3.7.0
seaborn>=0.12.0
tensorboard>=2.13.0
pillow>=9.5.0

# Evaluation Metrics
scipy>=1.10.0
scikit-learn>=1.2.0
pytorch-fid>=0.3.0

# API & Deployment - FastAPI
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
python-multipart>=0.0.6

# API & Deployment - Flask
flask>=2.3.0
flask-cors>=4.0.0
gunicorn>=21.2.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Code Quality
black>=23.7.0
flake8>=6.0.0
mypy>=1.4.0

# Logging
loguru>=0.7.0
